feat: implement modules block
